requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.8.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0

# AI and Video Generation
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
diffusers==0.24.0
accelerate==0.24.1
opencv-python==4.8.1.78
moviepy==1.0.3
imageio==2.33.0
imageio-ffmpeg==0.5.1
numpy==1.24.3

# Additional video processing
av==10.0.0
scikit-image==0.21.0
